Movie and Genre Models setup
One of the media formats for the application, movie, has been setup
with several values and validations.

Migration created and performed to establish the movie and genre aspects
within the database.

Created associations between the movie and genre models as a genre
has_many movies and a movie belongs_to a genre.