Game Model Setup and Genre Model Added Associations
A new game model with shared and individual attributes along with a private method
have been setup as part of the continued model setup process. Within the updated schema
are further attributes that will be added in due course to the game model.

Further has_many associations have been added on to the genre model to ensure
it is compatible with the newly added models.

The database schema has been updated via a migration following the creation of the new game model related
migration file.